Refactor authentication and company management: - Update user model to include author reference in company schema. - Modify company routes to enforce ownership checks. - Implement token verification middleware for protected routes. - Reorganize auth routes into a dedicated router. - Update package dependencies for bcrypt and jsonwebtoken.

+++ controllers/companyControllers.js
@@ -3,7 +3,7 @@ import Company from "../models/companyModel.js";
 // GET /companies/
 export const getCompanies = async (req, res) => {
   try {
-    const companies = await Company.find({});
+    const companies = await Company.find({ author: req.user._id });
     res.json(companies);
   } catch (e) {
     console.log(`Error at getCompanies: ${e}`);
@@ -14,7 +14,10 @@ export const getCompanies = async (req, res) => {
 // GET /companies/:id
 export const getCompany = async (req, res) => {
   try {
-    const company = await Company.find({ _id: req.params.id });
+    const company = await Company.findById(req.params.id);
+    if (!company.author.equals(req.user._id)) {
+      return res.status(403).send("Action not allowed!");
+    }
     res.json(company);
   } catch (e) {
     if (res.statusCode === 404) {
@@ -29,6 +32,7 @@ export const getCompany = async (req, res) => {
 // POST /companies/
 export const createCompany = async (req, res) => {
   try {
+    req.body.author = req.user._id;
     const createdCompany = await Company.create(req.body);
     res.status(201).json(createdCompany);
   } catch (e) {
@@ -39,6 +43,10 @@ export const createCompany = async (req, res) => {
 // DELETE /companies/:id
 export const deleteCompany = async (req, res) => {
   try {
+    const company = await Company.findById(req.params.id);
+    if (!company.author.equals(req.user._id)) {
+      return res.status(403).send("Action not allowed!");
+    }
     const deletedCompany = await Company.findByIdAndDelete(req.params.id);
     if (!deletedCompany) {
       res.status(404);
@@ -57,6 +65,10 @@ export const deleteCompany = async (req, res) => {
 // UPDATE /companies/:id
 export const updateCompany = async (req, res) => {
   try {
+    const company = await Company.findById(req.params.id);
+    if (!company.author.equals(req.user._id)) {
+      return res.status(403).send("Action not allowed!");
+    }
     const updatedCompany = await Company.findByIdAndUpdate(
       req.params.id,
       req.body,
